Code cleanup and parsed output
