+++ saviSimulation/savi/commandStation/CommandStationCore.java
@@ -1,24 +1,29 @@
 package savi.commandStation;
 
+import savi.commandStation.Telemetry.*;
+
 public class CommandStationCore {
 	
 	private CommandStationConnector connector; 	// connection to the rest of the simulation
 	private String id; 							// identifier to use in messages
 	
 	private TelemetryFetcher telemetryFetcher;	// class in a separate thread that generates messages for fetching telemetry
-	private FancyCommandStationGUI gui;
-	private TelemetryTracker telemetryTracker;	// Keeps track of the telemetry and organizes it
+	private CommandStationGUI gui;			// GUI for the command station
 
+
+	
+	/**
+	 * Constructor for the core of the command station
+	 * @param connector 	The connector to the rest of the simulation
+	 * @param ID			The agent ID for the command station (address for messages)
+	 */
 	public CommandStationCore(CommandStationConnector connector, String ID) {
 		// Set message ID and connecter parameters
 		this.id = ID;
 		this.connector = connector;
 		
-		// Setup the telemetry tracker
-		this.telemetryTracker = new TelemetryTracker();
-		
 		// Build the GUI (remove the parameters)
-		gui = new FancyCommandStationGUI(this);
+		gui = new CommandStationGUI(this);
 		
 		// Build the telemetry fetcher
 		this.telemetryFetcher = new TelemetryFetcher(this);
@@ -29,6 +34,13 @@ public class CommandStationCore {
 	}
 	
 	
+	/**
+	 * Send the message to the agents
+	 * 
+	 * @param destination - ID of the agent or BROADCAST for all agnets
+	 * @param messageType - message type, "achieve", "tell", or any other JASON supported message types
+	 * @param parameter - The message in AgentSpeak
+	 */
 	public synchronized void sendMessage(String destination, String messageType, String parameter) {
 
 		// this is a really crude message id!
@@ -36,7 +48,20 @@ public class CommandStationCore {
 		
 		// Build the message
 		String message = "<" + mid + "," + this.id + "," + messageType + "," + destination + "," + parameter + ">";
-			
+		
+		// Send the message
+		this.sendMessagePassthrough(message);
+	}
+	
+	
+	/**
+	 * Sends the message to the agents
+	 * 
+	 * *** Assuming that the messages are valid, socket receiver should call this method ***
+	 * 
+	 * @param	The message to be sent - MUST BE A VALID MESSAGE, NO ERROR CHECKING IN THIS METHOD
+	 */
+	public synchronized void sendMessagePassthrough(String message) {
 		// Send the message
 		this.connector.messageOut(message);
 	}
@@ -51,9 +76,11 @@ public class CommandStationCore {
 		
 		// *** SEND OVER SOCKET HERE *** //
 		
-		// Deal with receiving the message. Send to the GUI raw for now
-		gui.receiveMessage(msg);
-		telemetryTracker.update(msg);
+		// Parse the message and get a telemetryItem
+		TelemetryItem item = TelemetryItem.getTelemetryItem(msg);
+		
+		// Send to the GUI
+		gui.receiveMessage(msg, item.toString());
 	}
 }
